Use EuclideanVector int16_t instead of float

+++ src/imu/inertial_measurement_sensor.cpp
@@ -2,7 +2,7 @@
 
 namespace imu {
 
-auto InertialMeasurementSensor::Get(void) noexcept -> types::EuclideanVector<float> {
+auto InertialMeasurementSensor::Get(void) noexcept -> types::EuclideanVector<int16_t> {
   return sensor_values_;
 }
